requirements.txt
Flask==3.1.1
flask_cors==6.0.1
lxml==6.0.0
Pillow==11.3.0
reportlab==4.4.2
Requests==2.32.4
